fix: structure project, dir names, aliases, routing
