+++ src/App/pages/TestsPage/index.ts
@@ -1,0 +1,1 @@
+export { TestsPage } from "./TestsPage";
