+++ src/App/pages/index.ts
@@ -1,0 +1,2 @@
+export * from "./MainPage";
+export * from "./TestsPage";
