--- src/app/App.tsx
@@ -1,8 +1,0 @@
-import "./App.css";
-import { MainPage } from "../pages/main-page/main-page";
-
-function App() {
-  return <MainPage />;
-}
-
-export default App;
