+++ tsconfig.json
@@ -1,4 +1,14 @@
 {
+  "compilerOptions": {
+    "baseUrl": ".",
+    "paths": {
+      "src/*": ["src/*"],
+      "assets/*": ["src/assets/*"],
+      "pages/*": ["src/App/pages/*"],
+      "api/*": ["src/api/*"],
+      "styles/*": ["src/styles/*"]
+    }
+  },
   "files": [],
   "references": [
     { "path": "./tsconfig.app.json" },
